Adjusted video header writes.
SFML wasn't designed with unsigned long long to be written to a packet, strange. Upon adding this, bizzaire functionality arose which caused me to only keep time as it was consistent between Windows and Linux. Adding more than one 64 bit value will cause parsing issues.

+++ Source/Core/Core/SocketComm/OutputComm.hpp
@@ -103,7 +103,7 @@ namespace SocketComm
 		int SendUdpMessage(sf::Packet &packet);
 
 		/** Number of segments */
-		uint16_t mSegments = 1; 
+		uint8_t mSegments = 1; 
 		/** Number of frames that have passed */
 		uint32_t mFrameCount = 0;
 		/** Port to send on, either Slippi or Video */
